UpgradeNudge: move from my-sites to blocks
