+++ client/components/tinymce/plugins/simple-payments/dialog/index.jsx
@@ -43,7 +43,7 @@ import {
 } from 'state/simple-payments/product-list/actions';
 import { PLAN_PREMIUM, FEATURE_SIMPLE_PAYMENTS } from 'lib/plans/constants';
 import { hasFeature, getSitePlanSlug } from 'state/sites/plans/selectors';
-import UpgradeNudge from 'my-sites/upgrade-nudge';
+import UpgradeNudge from 'blocks/upgrade-nudge';
 import TrackComponentView from 'lib/analytics/track-component-view';
 import {
 	bumpStat,
