+++ client/my-sites/media-library/list-plan-upgrade-nudge.jsx
@@ -13,7 +13,7 @@ import { localize } from 'i18n-calypso';
  * Internal dependencies
  */
 import { FEATURE_VIDEO_UPLOADS, FEATURE_AUDIO_UPLOADS } from 'lib/plans/constants';
-import UpgradeNudge from 'my-sites/upgrade-nudge';
+import UpgradeNudge from 'blocks/upgrade-nudge';
 import ListPlanPromo from './list-plan-promo';
 
 function getTitle( filter, translate ) {
